+++ programs/clearing_house/src/state/state.rs
@@ -83,8 +83,8 @@ pub struct PriceDivergenceGuardRails {
 pub struct ValidityGuardRails {
     pub slots_before_stale_for_amm: i64,
     pub slots_before_stale_for_margin: i64,
-    pub confidence_interval_max_size: u128,
-    pub too_volatile_ratio: i128,
+    pub confidence_interval_max_size: u64,
+    pub too_volatile_ratio: i64,
 }
 
 #[derive(AnchorSerialize, AnchorDeserialize, Clone)]
